requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.3
google-cloud-build==3.29.0
google-cloud-run==0.10.14
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
responses==0.25.6
urllib3==2.3.0
